[COREVM-60] Code cleanup

+++ include/runtime/common.h
@@ -22,7 +22,7 @@ typedef uint64_t instr_oprd;
 typedef int32_t variable_key;
 
 
-typedef uint32_t gc_bitfield_t;
+typedef uint8_t gc_bitfield_t;
 
 
 const instr_addr NONESET_INSTR_ADDR = -1;
